Move headers to subdirectory
