+++ src/HeightGainTerminalCorrectionModel.cpp
@@ -1,6 +1,6 @@
-#include "../include/P2108.h"
-#include "../include/Errors.h"
-#include "../include/Consts.h"
+#include "../include/ITS.ITU.PSeries.P2108/P2108.h"
+#include "../include/ITS.ITU.PSeries.P2108//Errors.h"
+#include "../include/ITS.ITU.PSeries.P2108/Consts.h"
 
 /*=============================================================================
  |
